Fix compilation issues
