+++ functions/firestore-export/package.json
@@ -10,6 +10,7 @@
     },
     "dependencies": {
         "@google-cloud/firestore": "^4.11.1",
+        "firebase-admin": "^9.8.0",
         "firebase-functions": "^3.14.0"
     }
 }
