+++ lerna.json
@@ -2,6 +2,7 @@
   "lerna": "2.8.0",
   "packages": [
     "auth",
+    "auth/functions",
     "config",
     "database",
     "firebaseapp",
